requirements.txt
crewai[tools]
python-dotenv